Switched to building a JAR running Jetty instead of WAR
Building a JAR now instead of a WAR. This removes the requirement for
having an application server which used to startup pretty slowly
compared to Jetty. It also speeds up the docker push process if you
don't alter the dependent libraries as only the thin JAR needs to get
uploaded on a change (instead of the WAR which contained all libraries,
the libraries are now uploaded to the Docker image as a dedicated layer
which doesn't change regularly.